SC Data Imports for 3.17.2 (#2314)
Changed the way Imports of gamefile information is handled on Fleetyards.
The Import now relies on Exported scunpacked files on a cloud storage and saves the latest imported version into the Database instead of a file on disk. This will allow to add new gamefile Imports without the need of deployment.

To optimise the repository size for potential Heroku Review Apps and a future
Docker setup, seed files and sc_data imports should not be part of the
repository

Introduce new Imports::ScDataImport class which uses STI from the Import class
to track imports including its game version and status.